changed fragments implementation
now fragments are stored in two structs, one for just default fragments and one for fragments that can be changed (the second inherits from the firs).
items have the first struct, while stacks have the second.
the struct in stacks has its default fragments set when the item gets set

as of now it remains the concern of keeping the owning inventory component updated and to deal with adding / removing from the replicated subobject list