added consume item by definition and band-aid fixed an issue with OnItemCountChanged
OnItemCountChanged calls registerSlotChange, and if count is zero, passes the item as OldItem, which is problematic, since then the item would get unregistered from replicated sub-objects. thus the count would never replicate to client.
Fixed it by passing null in RegisterSlotChange. but will have to rework the system a bit to make it pretty